Refactor Unsafe related stuff in scala.concurrent.
