+++ src/library/scala/concurrent/util/Unsafe.java
@@ -6,22 +6,25 @@
 **                          |/                                          **
 \*                                                                      */
 
+package scala.concurrent.util;
 
-package scala.concurrent;
+
 
 import java.lang.reflect.Field;
 
-final class Unsafe {
+
+
+public final class Unsafe {
     public final static sun.misc.Unsafe instance;
     static {
         try {
             sun.misc.Unsafe found = null;
             for(Field field : sun.misc.Unsafe.class.getDeclaredFields()) {
-              if (field.getType() == sun.misc.Unsafe.class) {
-                field.setAccessible(true);
-                found = (sun.misc.Unsafe) field.get(null);
-                break;
-              }
+		if (field.getType() == sun.misc.Unsafe.class) {
+		    field.setAccessible(true);
+		    found = (sun.misc.Unsafe) field.get(null);
+		    break;
+		}
             }
             if (found == null) throw new IllegalStateException("Can't find instance of sun.misc.Unsafe");
             else instance = found;
